added convert to jsarray and jsobject functions

+++ native/src/knn.rs
@@ -2,7 +2,7 @@ use rayon::prelude::*;
 use distance_computation::{compute_distance, euclidean};
 
 
-pub fn knn(target: &Vec<f64>, k: u32, neighbors: &Vec<(String, Vec<f64>)>) 
+pub fn knn(target: &Vec<f64>, k: f64, neighbors: &Vec<(String, Vec<f64>)>) 
   -> Vec<(String, f64)> {
     let distances = get_distances(target, neighbors);
     get_k(distances, k)
@@ -17,7 +17,7 @@ fn get_distances(target: &Vec<f64>, neighbors: &Vec<(String, Vec<f64>)>)
 }
 
 
-fn get_k(mut dists: Vec<(String, f64)>, k: u32) -> Vec<(String, f64)> {
+fn get_k(mut dists: Vec<(String, f64)>, k: f64) -> Vec<(String, f64)> {
   dists.sort_by(|(_, a), (_, b)| a.partial_cmp(b).unwrap());
   dists.iter()
        .take(k as usize)
